Remove one-off interface
Also organize file structure to reflect type of underlying data source with which code is concerned.
