+++ src/jplyr.jl
@@ -3,11 +3,8 @@ using Reexport
 include("AbstractTables/AbstractTables.jl")
 @reexport using .AbstractTables
 
-export @query,
-    @select,
-    @filter,
-    @groupby,
-    @summarize
+export  @query,
+        @qcollect
     # following are exported only for test purposes
 
 # Graph interface
@@ -18,15 +15,15 @@ include("query.jl")
 include("abstracttable/assignment_expr_ops.jl")
 include("abstracttable/sym_analysis.jl")
 
-# One-off interface
-include("select.jl")
-include("filter.jl")
-include("groupby.jl")
-include("summarize.jl")
-
-# Execution
+# Generic graph execution
 include("collect.jl")
-include("generic.jl")
-include("utils.jl")
+
+# Graph execution over AbstractTables
+include("abstracttable/collect.jl")
+include("abstracttable/generic.jl")
+include("abstracttable/utils.jl")
+include("abstracttable/select.jl")
+include("abstracttable/filter.jl")
+include("abstracttable/summarize.jl")
 
 end # module jplyr
