+++ test/runtests.jl
@@ -3,10 +3,7 @@ quiet = length(ARGS) > 0 && ARGS[1] == "-q"
 anyerrors = false
 
 my_tests = (
-    "select.jl",
-    "filter.jl",
-    "summarize.jl",
-    "query.jl"
+    "query.jl",
 )
 
 println("Running tests:")
